Initial state of refactoring
- Separate functional to packets: containerinfo, k8s, ssh
