--- internal/nsenter/containerinfo.go
@@ -1,138 +1,0 @@
-package nsenter
-
-import (
-	"context"
-	"fmt"
-	"path"
-	"strings"
-
-	"github.com/pabateman/kubectl-nsenter/internal/config"
-
-	"github.com/pkg/errors"
-	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/tools/clientcmd"
-)
-
-type ContainerInfo struct {
-	NodeName         string
-	NodeIP           string
-	ContainerID      string
-	ContainerRuntime string
-}
-
-func NewPodSpec(name, namespace string, kubeClient *kubernetes.Clientset) (*v1.Pod, error) {
-	return kubeClient.CoreV1().Pods(namespace).Get(context.TODO(), name, metav1.GetOptions{})
-}
-
-func PodInitialized(pod *v1.Pod) bool {
-	for _, condition := range pod.Status.Conditions {
-		if condition.Type == v1.PodInitialized {
-			return condition.Status == v1.ConditionTrue
-		}
-	}
-	return false
-}
-
-func GetInitializingContainerStatus(pod *v1.Pod) (*v1.ContainerStatus, error) {
-	for _, status := range pod.Status.InitContainerStatuses {
-		if status.State.Running != nil {
-			return &status, nil
-		}
-		continue
-	}
-	return nil, fmt.Errorf("none of initContainers is running")
-}
-
-func GetContainerStatus(pod *v1.Pod, cfg config.Config) (*v1.ContainerStatus, error) {
-	container := cfg.Container
-
-	containerStatuses := make([]v1.ContainerStatus, 0)
-	containerStatuses = append(containerStatuses, pod.Status.ContainerStatuses...)
-	containerStatuses = append(containerStatuses, pod.Status.EphemeralContainerStatuses...)
-
-	if container != "" {
-		for _, status := range containerStatuses {
-			if status.Name == container {
-				if status.State.Running != nil {
-					return &status, nil
-				}
-				return nil, fmt.Errorf("specified container is not running")
-			}
-			continue
-		}
-		return nil, fmt.Errorf("can't find specified container")
-
-	}
-	for _, status := range containerStatuses {
-		if status.State.Running != nil {
-			return &status, nil
-		}
-		continue
-	}
-
-	return nil, fmt.Errorf("pod %v has no running containers", pod.Name)
-}
-
-func GetClientSet(cfg config.Config) (*kubernetes.Clientset, string, error) {
-	config := clientcmd.NewNonInteractiveDeferredLoadingClientConfig(
-		&clientcmd.ClientConfigLoadingRules{Precedence: strings.Split(cfg.KubeConfig, ":")},
-		&clientcmd.ConfigOverrides{
-			CurrentContext: cfg.KubeContext,
-		})
-
-	clientConfig, err := config.ClientConfig()
-	if err != nil {
-		return nil, "", errors.Wrap(err, "can't build config")
-	}
-
-	clientset, err := kubernetes.NewForConfig(clientConfig)
-	if err != nil {
-		return nil, "", errors.Wrap(err, "can't build client")
-	}
-
-	namespace, _, err := config.Namespace()
-	if err != nil {
-		return nil, "", errors.Wrap(err, "can't get current namespace")
-	}
-
-	return clientset, namespace, nil
-}
-
-func GetContainerInfo(cfg config.Config) (*ContainerInfo, error) {
-	kubeClient, namespace, err := GetClientSet(cfg)
-	if err != nil {
-		return nil, errors.Wrap(err, "can't build client")
-	}
-
-	if ns := cfg.Namespace; ns != "" {
-		namespace = ns
-	}
-
-	podSpec, err := NewPodSpec(cfg.PodName, namespace, kubeClient)
-	if err != nil {
-		return nil, errors.Wrap(err, "can't get pod spec")
-	}
-
-	var containerStatus *v1.ContainerStatus
-
-	if !PodInitialized(podSpec) {
-		containerStatus, err = GetInitializingContainerStatus(podSpec)
-		if err != nil {
-			return nil, fmt.Errorf("pod is initializing: %v", err)
-		}
-	} else {
-		containerStatus, err = GetContainerStatus(podSpec, cfg)
-		if err != nil {
-			return nil, fmt.Errorf("can't get container status: %v", err)
-		}
-	}
-
-	return &ContainerInfo{
-		NodeName:         podSpec.Spec.NodeName,
-		NodeIP:           podSpec.Status.HostIP,
-		ContainerID:      path.Base(containerStatus.ContainerID),
-		ContainerRuntime: strings.Split(containerStatus.ContainerID, ":")[0],
-	}, nil
-}
